remove native implementation and always use JS
V8 has been continuously improved and there is no significant performance difference between the native implementation and the JS one.
