+++ .npmignore
@@ -1,4 +1,3 @@
-build/
 test/
 appveyor.yml
 .*
