+++ .travis.yml
@@ -1,4 +1,3 @@
-sudo: false
 language: node_js
 node_js:
   - "10"
@@ -7,16 +6,3 @@ node_js:
 os:
   - linux
   - osx
-before_deploy:
-  - ARCHIVE_NAME="${TRAVIS_TAG:-latest}-$TRAVIS_OS_NAME-$(uname -m).tar"
-  - npm run prebuild
-  - tar --create --verbose --file="$ARCHIVE_NAME" --directory "$TRAVIS_BUILD_DIR/prebuilds" .
-deploy:
-  provider: releases
-  api_key:
-    secure: VCE2AgIOeQFDHWw9g+pRmh8KibbIoTstKlUsVVHrWrICTCYdoSA7ivvhy40MRi8uJwK+cMbsf3auSLYJ6siUdcm46S+5IiU65ZBCGhzhaFOJPmVdzbqK2TNM3dEVFePsnLq1Oki5YzbdjZg46AIZFrAhCmuFahicnDcY9cN2/fg=
-  file: "$ARCHIVE_NAME"
-  skip_cleanup: true
-  on:
-    tags: true
-    node: "10"
