+++ appveyor.yml
@@ -17,17 +17,3 @@ test_script:
   - node --version
   - npm --version
   - npm test
-after_test:
-  - ps: If ($env:nodejs_version -eq "10") { npm run prebuild }
-  - ps: $env:ARTIFACT_NAME_PREFIX = if (Test-Path env:APPVEYOR_REPO_TAG_NAME) { $env:APPVEYOR_REPO_TAG_NAME } else { 'latest' }
-artifacts:
-  - path: prebuilds
-    name: $(ARTIFACT_NAME_PREFIX)-win-$(PLATFORM)
-deploy:
-  - provider: GitHub
-    artifact: /.*\.zip/
-    auth_token:
-      secure: CkWRNlzH+wmCa3+TRydWfvmSH7tevR81aAypEhXTR9ka5q0Ja8lAIPDwtPb+Ux9l
-    on:
-      appveyor_repo_tag: true
-      nodejs_version: "10"
