--- binding.gyp
@@ -1,9 +1,0 @@
-{
-  'targets': [
-    {
-      'target_name': 'validation',
-      'sources': ['src/validation.c'],
-      'cflags': ['-std=c99']
-    }
-  ]
-}
